incorporated implemented store to game
