+++ packages/app/src/app/race/_components/multiplayer/waiting-screen.tsx
@@ -1,4 +1,5 @@
 "use client";
+import { RACE_STATUS, type RaceStatus } from "@code-racer/wss/src/consts";
 
 import React from "react";
 
@@ -9,11 +10,11 @@ type WaitingScreenProps = {
   roomID: string;
   amountOfPlayers: number;
   IS_PLAYER_CURRENT_USER: boolean;
-  handleResetGame: () => void;
+  handleChangeGameState: (_status: RaceStatus) => void;
 };
 
 const WaitingScreen: React.FC<WaitingScreenProps> = React.memo(
-  ({ roomID, amountOfPlayers, IS_PLAYER_CURRENT_USER, handleResetGame }) => {
+  ({ roomID, amountOfPlayers, IS_PLAYER_CURRENT_USER, handleChangeGameState }) => {
     return (
       <React.Fragment>
         <div className="mb-12">
@@ -29,7 +30,7 @@ const WaitingScreen: React.FC<WaitingScreenProps> = React.memo(
             <span>You are the room owner of this room.</span>
             <Button
               className="mt-2 w-full"
-              onClick={handleResetGame}
+              onClick={() => handleChangeGameState(RACE_STATUS.COUNTDOWN)}
               disabled={amountOfPlayers < 2}
             >
               {amountOfPlayers > 1
